Add --stash flag to `pueue add`
If the queue is empty and the user adds a task, the deamon will start it
at its convince. This is faster than most people can type. However, if
the user wants it to be stashed, they missed their opportunity. By
adding a `--stash` flag, the user can specify the task should be created
in the stashed state allowing them to `pueue enqueue` them at a future
time.

Since it does not make sense to specify both `--immediate` and
`--stash`, I marked the options as conflicting.